Create Spontaneous-Casting-Cure
IF YOU ARE A CLERIC (with this ability) YOU MUST: Add a Special Ability to your character sheet in the special abilities section. Call it Spontaneous-Casting-Cure in the description field (the long one accessible by the GEAR icon) put the text: AbilityScripts;Cleric. In the Curr field/column, put the letter u. To spontaneously cast a cure spell, click Special-Abilities and then select the Spontaneous-Casting-Cure option.